pyproject.toml
[tool.poetry]
name = "vdeploy"
version = "0.1.0"
description = ""
authors = ["Ada Bohm <ada@kreatrix.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
vast-ai-api = "1.0"
vastai = "0.2"
pyserde = { version = "0.16", extras = ["toml"]}
prettytable = "3.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
